pyproject.toml
[project]
name = "curriculo"
version = "0.1.0"
description = ""
authors = [
    {name = "Lorenzo"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-api-python-client (>=2.166.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
